+++ method_to_swagger_yaml/lib/src/entity/section/paths_section/body/request_body_section.dart
@@ -20,6 +20,7 @@ class RequestBodySection {
         methodName: methodName,
         variableName: parameter.name,
         returnType: parameter.type,
+        element: parameter,
       );
 
       ComponentSection.add(componentSection);