+++ sample_pj/lib/repository/object/skip_target.g.dart
@@ -1,0 +1,19 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'skip_target.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$SkipTargetImpl _$$SkipTargetImplFromJson(Map<String, dynamic> json) =>
+    _$SkipTargetImpl(
+      skipTarget: json['skipTarget'] as String,
+      ignoreTarget: json['ignoreTarget'] as String,
+    );
+
+Map<String, dynamic> _$$SkipTargetImplToJson(_$SkipTargetImpl instance) =>
+    <String, dynamic>{
+      'skipTarget': instance.skipTarget,
+      'ignoreTarget': instance.ignoreTarget,
+    };