+++ method_to_swagger_yaml_annotation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+- add IgnoreFieldInYaml annotation
+
 ## 0.0.2
 
 - add useToString option in RequestParameter